[RIKU-8] refactor: RunList 컴포넌트 MainTabLayout 내부의 자식 router로 통합 및 TabNavigationUI-detail 수정
